backend/requirements.txt
Django>=5.1,<5.2
djangorestframework>=3.15,<3.16
django-cors-headers>=4.6,<4.7
PyMuPDF>=1.27,<1.28
Pillow>=10.0,<12.0
